pyproject.toml
[tool.poetry]
name = "seed-salesforce"
version = "0.1.0"
description = "Package for connecting SEED data to Salesforce"
authors = ["Nicholas Long <nicholas.long@nrel.gov>"]
license = "BSD-4-Clause"

readme = "README.md"

homepage = "https://seed-platform.org"
repository = "https://github.com/seed-platform/tbd"
documentation = "https://seed-platform.org"
keywords = ["SEED Platform", "Building Benchmarking", "Salesforce"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
packages = [
    { include = "seed_salesforce"}
]

[tool.poetry.dependencies]
python = ">=3.7.1, <3.10"
python-dateutil = "> 0"
# py-seed = "^0.2.1"
py-seed = { path = "../py-seed", develop = true}
# careful using github with poetry. Often when calling poetry update, it will not pull down the latest commits and normally requires a removal/recreation of the virtualenv
# py-seed = { git = "https://github.com/seed-platform/py-seed.git", branch = "add-seed-helpers" }

simple-salesforce = "^1.11.6"
usaddress = ">0"

[tool.poetry.dev-dependencies]
autopep8 = "1.5.4"
flake8 = "3.8.3"
mypy = "^0.950"
pre-commit = "2.19.0"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
tox = "3.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
